third_party/blink/web_tests/fast/forms/checkValidity-on-cloned-node-expected.txt

Test checkValidity correctness on input select and textarea element after cloning node .

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.checkValidity() is true
PASS clonedInput.checkValidity() is true
PASS clonedInputDeep.checkValidity() is true
PASS input.checkValidity() is false
PASS clonedInputNext.checkValidity() is false
PASS select.checkValidity() is true
FAIL clonedSelect.checkValidity() should be true. Was false.
PASS clonedSelectDeep.checkValidity() is true
PASS textarea.checkValidity() is true
PASS clonedTextArea.checkValidity() is true
PASS clonedTextAreaDeep.checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
